+++ statewide-death-profiles/update.sh
@@ -1,14 +1,26 @@
 
-uuid1="7a456555-87b9-4830-817c-72d72e628745"
-uuid2="9964e86b-8796-4f59-8880-c5f736763e7b"
+if [ "$1" = "--help" ] || [ "$1" = "-h" ]; then
+    echo ""
+    echo "usage: bash update.sh [ --no-fetch ]"
+    echo ""
+    exit 
+fi
 
-/bin/rm -f *zip *zip.[0-9]*
+if [ "$1" != "--no-fetch" ]; then
 
-wget "https://data.chhs.ca.gov/dataset/$uuid1/resource/$uuid2/download/statewide-death-profiles-8eleum.zip"
+    uuid1="7a456555-87b9-4830-817c-72d72e628745"
+    uuid2="9964e86b-8796-4f59-8880-c5f736763e7b"
 
-/bin/rm -f *csv *json
+    /bin/rm -f *zip *zip.[0-9]*
 
-unzip statewide-death-profiles-8eleum.zip
+    echo "fetching..."
 
-SQLALCHEMY_SILENCE_UBER_WARNING=1 ./venv/bin/python deaths.py
+    wget -q "https://data.chhs.ca.gov/dataset/$uuid1/resource/$uuid2/download/statewide-death-profiles-8eleum.zip"
+
+    /bin/rm -f *csv *json
+
+    unzip statewide-death-profiles-8eleum.zip
+fi
+
+SQLALCHEMY_SILENCE_UBER_WARNING=1 ./.venv/bin/python deaths.py
